Check out kcapi-sys as a Submodule
This crate provides a Rusty wrapper to kcapi-sys, and therefore
in order to allow easy test and build (and until the kcapi-sys
crate is published) we need to add the kcapi-sys repository as
a submodule of this repository.

This change adds the kcapi-sys repository as a submodule of this
repository.