lib: Create the KcapiError, KcapiHandle types
We need data structures to hold and rustify the errors which the
libkcapi may throw at us. Additionally, libkcapi defines the
opaque struct kcapi_handle, which we need to allocate and implement
correctly in our bindings for this library.

Therefore, This change:

* Defines the KcapiError type to hold errors from libkcapi
* Defines the KcapiHandle struct and implements a constructor for
  it in order to allocate it in a rusty way.

Note: We need to redefine the kcapi_handle in this crate as the
member _unused is private in kcapi-sys crate. This is a hack for
now which we will resolve later.