md: Add the Rusty API for the kcapi_md_* APIs
We have implemented a fairly rusty API for the kcapi_md_* calls.
This change commits these APIs so that they officially are provided
by the kcapi crate.

This change:

* Creates a new module 'md' to hold this API
* Adds a number of Rusty bindings for the kcapi_md_* calls:

    kcapi::md::init()
    kcapi::md::update()
    kcapi::md::digest_final()
    kcapi::md::digest()
    kcapi::md::digestsize()
    kcapi::md::destroy()

In addition to these, the kcapi_md_sha* calls as well as the
kcapi_md_hmac_* calls have also been provided for. This is
the first set of APIs for the kcapi crate.