skcipher: Add Rusty API for Symmetric Ciphers
The next step to writing the Rusty bindings for libkcapi involves
adding support for symmetric ciphers. We have implemented a good
API for the various kcapi_cipher_* calls from libkcapi.

This change adds the high-level rust bindings for these calls:

	kcapi::skcipher::alg_init()
	kcapi::skcipher::alg_setkey()
	kcapi::skcipher::alg_encrypt()
	kcapi::skcipher::alg_decrypt()

This also provides direct support for AES CBC and CTR enc/decryption
as well as provides a one-shot API for en/decryption for any algorithm
defined in /proc/crypto.