test: Add Tests for the skcipher Module
Since we have implemented symmetric key support for the kcapi module,
it becomes necessary to provide tests for the same to maintain
good quality of code. This change adds these tests.

With these tests, we can unit-test all of the skcipher APIs. The
next step will be to provide Asynchronous I/O support.