rng: Implement a Rusty API for kcapi_rng_*
We have implemented a fairly rusty API for the various kcapi_rng_*
calls from libkcapi. These APIs are provided in the rng module of
the kcapi crate.

This change implements the high-level rust bindings for:

	- kcapi::rng::init()
	- kcapi::rng::seed()
	- kcapi::rng::generate()
	- kcapi::rng::get_bytes()

With this complete random number generation support from libkcapi
is provided to the kcapi crate