aead: Implement a Rusty API for kcapi_aead_*
We have implemented a fairly rusty API for the various kcapi_aead_*
calls from libkcapi. These APIs are provided in the aead module of
the kcapi crate.

This change implements the high-level rust bindings for:

	- kcapi::aead::alg_init()
	- kcapi::aead::encrypt()
	- kcapi::aead::decrypt()

Other functions are also implemented which allow the caller to
control the various interactions with the aead subsystem directly.
However, these need to be improved to be made more fine-grained.