md: Refactor Code to Use KcapiHash Implementation
Currently the md code depends a lot on the KcapiHandle struct.
It is passed around different calls in order to maintain context
between them. The same context can be implemented by a public
KcapiHash struct with the init, update, finalize et. al.
methods being folded into it's impl.

This change:

* Removes the alg_* APIs for kcapi::md
* Creates a KcapiHash struct to hold the context for md operations
* Folds the init, update, finalize, setkey etc. methods into
  this struct's impl.