skcipher: Refactor Code to Use KcapiSKCipher impl
Currently the skcipher code depends on KcapiHandle struct to
maintain context between a number of calls such as init, setkey,
encrypt, and decrypt as well as stream and ayncio operations for
the same. The same context can be implemented with a public
KcapiSKCipher type with this functionality folded into it's impl.

This change:

* Removes the alg_* type calls in skcipher.rs
* Creates a KcapiSKCipher type to hold context for cipher operations
* Implements the encrypt, decrypt, init, setkey, et. al. methods
  for the KcapiSKCipher Type