md: Remove the flags Argument from KcapiHash
The kcapi_hash_init() API from libkcapi provides a flags argument
to denote whether an asynchronous IO operation is required, but
no API to perform an AIO hash is provided. Therefore, we can
safely eliminate this flag from the initialization function
for the KcapiHash Type.

This change removes the 'flags' argument from the KcapiHash type.