kdf: Remove the flags argument from KcapiKDF
The kcapi_md_init() API from kcapi-sys requires a flags argument
to denote whether an asynchronous IO operation is required, but
no API to perform an AIO hash (or KDF) is provided. Therefore, we
can safely eliminate this flag from the initialization function of
the KcapiKDF type.

This change removes the 'flags' argument from the new() funcion
for the KcapiKDF type.