aead: Implement the KcapiAEADData Type
The we need a better way to provide input to the en/decrypt
APIs from the `kcapi::aead` module. For this we have implemented
a KcapiAEADData type, which has the following fields:

	KcapiAEADData {
	    data: Vec<u8>	// The data to be en/decrypted
	    assocdata: Vec<u8>	// The associated data
	    tag: Vec<u8>	// The authentication tag
	    tagsize: usize	// The size of the tag
	}

This type also implements the following methods:

	- set_{data,assocdata,tag}()
	- get_{data,assocdata,tag}()
	- {tag,data,assoc}len()
	- new_enc() // initialize data to be encrypted
	- new_dec() // initialize data to be decrypted

This change makes the necessary updates to the aead module and
it's tests to accomodate this new type.