Implement Drop Trait for Various Algorithm Handles
Currently, the kcapi_*_destroy() calls are not called whenever the
Kcapi{Hash,SKCipher,KDF,AKCipher,AEAD} types go out of scope. This
results in the associated kcapi_handle never getting freed,
resulting in memory leaks and also the persistence of the AF_ALG
sockets created by these handles as open file descriptors.

This causes issues when we attempt to scale the API calls and may
also result in a distributed denial-of-service condition when a
large number of these calls are made in parallel.

This change:

* Implements the Drop trait for:

	- KcapiHash
	- KcapiKDF
	- KcapiSKCipher
	- KcapiAEAD
	- KcapiAKCipher

Whenever any instances of these types go out of scope, the
associated kcapi_handle is freed using kcapi_sys::kcapi_*_destroy().