skcipher: Add Tests for AES-ECB
We did not have a method to test the operation in AES-ECB mode.
Now that 0-length IVs are allowed in KcapiSKCipher, it makes sense
that we implement a few tests which check the operation of our
bindings in AES-ECB mode.

This change adds a simple known-answer test to check the operation
of the KcapiSKCipher in AES-ECB mode.