Change KcapiError 'code' to i32
The field 'code' in type KcapiError is an i64 which is not standard
for the underlying libkcapi. We need this to be an i32 so that we
can pass up errno style codes to any consumers of this error.

In the future, the 'code' and 'message' will be public fields, so
that the error returned by the kernel can be appropriately
interpreted by consumers.

This change makes treewide updates to accomodate this change.