Introduce the VMSplice Trait
The libkcapi provides the capability of using VM splicing to avoid
copying data to the kernel. Therefore, kcapi APIs which allow the
setting of VM splice access to the kernel can make use of this
trait in order to provide the {get,set}_max_splicesize APIs as
a part of their implementation.

This change:

* Provides the VMSplice trait which can be implemented to wrap:
	- kcapi_get_maxsplicesize
	- kcapi_set_maxsplicesize