build: Add Build Features {local,vendored}-kcapi
Since kcapi-sys v1.4.0 we have the ability to build kcapi-sys
against a locally built libkcapi. We can extend this feature to
the kcapi crate and allow consumers of this crate to choose
whether to compile against the vendored library or a local copy
of the library. For now, to prevent any breaking changes,
vendored is the default.

This change:

* Adds a local-kcapi build feature
* Adds a vendored-kcapi build feature